functions/requirements.txt
azure-functions==1.20.0
azure-storage-blob==12.22.0
python-multipart==0.0.9
